Fix wrong username issue and better PR merge info
Previously usernames for issues and pull requests were always the name
of the user associated with the issue or PR (i.e. the user who opened
it). Now the displayed name is the username refers to the sender of the
event, meaning the one who actually performed the open, close, etc.

Also, now pull request close messages display if the pull request was
merged or not.